requirements.txt
beautifulsoup4==4.8.2
gevent==1.4.0
flask==1.1.1
bootstrap-flask==1.2.0
certifi==2019.11.28
chardet==3.0.4
gitdb2==2.0.6
GitPython==3.0.5
idna==2.8
PyYAML==5.3
requests==2.22.0
six==1.14.0
smmap2==2.0.5
thrift==0.13.0
tornado==6.0.3
urllib3==1.25.8
yuuki-core==6.5
